pyproject.toml
[project]
name = "assignment1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "blessed>=1.22.0",
    "fastapi[standard]>=0.121.2",
    "jupyter>=1.1.1",
    "pysqlite3>=0.5.4",
    "pytest>=8.4.2",
    "python-dotenv>=1.2.1",
]
